report custom events during ad session
