+++ src/ts/helper/AdBreakHelper.ts
